Corrected build errors and made intakespeed command more versatile

+++ src/main/java/frc/robot/constants/IntakeConstants.java
@@ -3,5 +3,5 @@ package frc.robot.constants;
 public class IntakeConstants {
 	public static final int motorId = 5;
 	public static final int sensorId = 2;
-	public static final int maxMotorSpeed = 0.80;
+	public static final double maxMotorSpeed = 0.80;
 }
